% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_venn_diagram.R
\name{create_venn_diagram}
\alias{create_venn_diagram}
\title{Create Venn Diagrams for Multiview Data}
\usage{
create_venn_diagram(M, areas, n_intersections, category, mainDir, subDir)
}
\arguments{
\item{M}{Number of views.}

\item{areas}{A vector of areas for each set.}

\item{n_intersections}{A matrix of intersections.}

\item{category}{A vector of category labels.}

\item{mainDir}{Main directory for saving plots.}

\item{subDir}{Subdirectory for saving plots.}
}
\description{
This function generates Venn diagrams based on the number of views.
}
