% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Friedman_function.R
\name{Friedman_function}
\alias{Friedman_function}
\title{Generate Non-Linear Effects Using Friedman Function}
\usage{
Friedman_function(x)
}
\arguments{
\item{x}{Matrix. A matrix where only the first 5 columns are used to generate non-linear effects.}
}
\value{
Numeric vector. The generated non-linear effects.
}
\description{
This function generates non-linear effects based on the Friedman function.
}
\examples{
x <- matrix(runif(50), ncol = 5)
y <- f(x)
}
