% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_simmba.R
\name{gen_simmba}
\alias{gen_simmba}
\title{This function generates simulated multi-omics data.}
\usage{
gen_simmba(
  nsample,
  snr = 1,
  p.train = 0.7,
  de.prob = rep(0.1, 3),
  de.downProb = rep(0.5, 3),
  de.facLoc = rep(1, 3),
  de.facScale = rep(0.4, 3),
  ygen.mode = "Friedman",
  nrep = 100,
  seed = 1234
)
}
\arguments{
\item{nsample}{Sample size}

\item{snr}{Signal to noise ratio}

\item{p.train}{Train-test split ratio}

\item{de.prob}{DE probability across all modalities}

\item{de.downProb}{Down-regulation probability}

\item{de.facLoc}{DE factor location}

\item{de.facScale}{DE factor scale}

\item{ygen.mode}{Y generation mode}

\item{nrep}{Number of repetitions}

\item{seed}{Random seed}
}
\value{
List containing training and testing datasets
}
\description{
Generates simulated multi-omics datasets with specified parameters,
including sample size, signal-to-noise ratio, DE probabilities,
and response variable generation mode. It also splits the data into training and testing sets.
}
\examples{
simulated_data <- gen_simmba(nsample = 100)
}
