% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBlockDim_function.R
\name{getBlocksDim}
\alias{getBlocksDim}
\title{Generate Block Dimensions Based on Dirichlet Distribution}
\usage{
getBlocksDim(K, p, alpha = 2)
}
\arguments{
\item{K}{Integer. The number of blocks.}

\item{p}{Integer. The total number of elements to be distributed among the blocks.}

\item{alpha}{Numeric. The concentration parameter for the Dirichlet distribution. Default is 2.}
}
\value{
Integer vector. A sorted vector of block sizes in descending order.
}
\description{
This function generates block dimensions based on the Dirichlet distribution, ensuring no block has zero size.
}
\examples{
# Generate block dimensions for 5 blocks and 100 elements
getBlocksDim(K = 5, p = 100)
}
