% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotmat_function.R
\name{plotmat}
\alias{plotmat}
\title{Plot a Matrix with Specified Settings}
\usage{
plotmat(
  mat_m,
  n_colors = 256,
  lab_col = NULL,
  is_cor = FALSE,
  folder = "~/Desktop/",
  filename = "prova"
)
}
\arguments{
\item{mat_m}{Matrix. The matrix to be plotted.}

\item{n_colors}{Integer. The number of colors to use in the plot. Default is 256.}

\item{lab_col}{Character. Optional label for the color legend. Default is NULL.}

\item{is_cor}{Logical. If TRUE, the matrix is treated as a correlation matrix with values ranging from -1 to 1. Default is FALSE.}

\item{folder}{Character. The folder to save the plot file. Default is '~/Desktop/'.}

\item{filename}{Character. The filename for the plot file (without extension). Default is 'prova'.}
}
\value{
None. The function saves the plot as a PNG file.
}
\description{
This function plots a matrix with specified settings, creating an image file.
}
\examples{
# Example matrix
mat <- matrix(runif(100), nrow = 10)

# Plot the matrix
plotmat(mat, n_colors = 256, lab_col = "Intensity", is_cor = FALSE, folder = tempdir(), filename = "matrix_plot")
}
